+++ include/Engine/ResourceManager.hpp
@@ -8,6 +8,7 @@
 #define RESOURCEMANAGER_H
 
 #include <map>
+#include <glm/glm.hpp>
 
 namespace Engine::Rendering {
   class Shader;
@@ -40,12 +41,21 @@ namespace Engine {
         const std::string &vShaderFile, const std::string &fShaderFile, const std::string &gShaderFile,
         const std::string &name
       );
+
       /**
        * Finds and returns the shader with the given name
        * @param name The name of the shader
        * @note Calling this method will log an error if the shader with the given name does not exist
        */
       static std::shared_ptr<Shader> GetShader(const std::string &name);
+
+      /**
+       * Finds and returns the shader with the given od
+       * @param id The id of the shader
+       * @note Calling this method will log an error if the shader with the given name does not exist
+       */
+      static std::shared_ptr<Shader> GetShaderById(uint id);
+
       /**
        * Loads a texture to memory
        * @param filePath The texture file path
@@ -56,25 +66,47 @@ namespace Engine {
        * - the file can't be read or is empty
        * @return The loaded texture
        */
-      static std::shared_ptr<Texture> LoadTexture(const std::string &filePath, bool alpha, const std::string &name);
+      static std::shared_ptr<Texture> LoadTexture2D(const std::string &filePath, bool alpha, const std::string &name);
+
       /**
        * Finds and returns the texture with the given name
-       * @param name
+       * @param name The name of the texture
        * @note Calling this method will log an error if the texture with the given name does not exist
        */
-      static std::shared_ptr<Texture> GetTexture(const std::string &name);
+      static std::shared_ptr<Texture> GetTexture2D(const std::string &name);
+
+      /**
+       * Finds and returns the texture with the given name
+       * @param id The id of the texture
+       * @note Calling this method will log an error if the texture with the given name does not exist
+       */
+      static std::shared_ptr<Texture> GetTexture2DById(uint id);
+
       /**
        * Creates a sprite using the texture with the given name
        * @param textureName The name of the texture to base this sprite on
        * @note Log's an error if no texture with the given name was found
        */
-      static std::shared_ptr<Sprite> CreateSprite(const std::string &textureName);
+      static std::shared_ptr<Sprite> CreateSpriteFromTexture(const std::string &textureName);
+
+      /**
+       * Creates a sprite using the texture with the given name
+       * @param spriteName The name of the new sprite
+       * @param textureName The name of the texture to base this sprite on
+       * @param rect The area of the texture to use to represent the sprite
+       * @note Log's an error if no texture with the given name was found
+       */
+      static std::shared_ptr<Sprite> CreateSprite(
+        const std::string &spriteName, const std::string &textureName, const glm::vec4 &rect = glm::vec4(0, 0, 1, 1)
+      );
+
       /**
        * Finds or create the sprite with the given name and returns it
        * @param name The name of the sprite, must be the same as the loaded textures name
        * @note Log's an error if no texture with the given name was found
        */
       static std::shared_ptr<Sprite> GetSprite(const std::string &name);
+
       /**
        * Loads a texture to memory
        * @param filePath The texture file path
@@ -85,9 +117,10 @@ namespace Engine {
        * - the file can't be read or is empty
        * @return The loaded texture
        */
-      static std::pair<std::shared_ptr<Texture>, std::shared_ptr<Sprite>> LoadTextureAndSprite(
+      static std::pair<std::shared_ptr<Texture>, std::shared_ptr<Sprite>> LoadTexture2DAndSprite(
         const std::string &filePath, bool alpha, const std::string &name
       );
+
       /// Deletes all the loaded textures and shaders
       static void Clear();
     private: