+++ src/Engine2D/ParticleSystem/ParticleSystemRenderer2D.cpp
@@ -10,17 +10,33 @@
 
 namespace Engine2D {
   std::vector<std::shared_ptr<ParticleSystem2D>> ParticleSystemRenderer2D::particleSystems;
+  std::vector<std::shared_ptr<ParticleSystem2D>> ParticleSystemRenderer2D::particleSystemsToAdd;
+  std::vector<std::shared_ptr<ParticleSystem2D>> ParticleSystemRenderer2D::particleSystemsToRemove;
+  #if MULTI_THREAD
   std::mutex ParticleSystemRenderer2D::particleSystemMutex;
+  #endif
 
   void ParticleSystemRenderer2D::update() {
+    #if MULTI_THREAD
     std::lock_guard lock(particleSystemMutex);
+    #endif
 
     for (const auto &particleSystem: particleSystems)
       particleSystem->update();
   }
 
   void ParticleSystemRenderer2D::render() {
+    #if MULTI_THREAD
     std::lock_guard lock(particleSystemMutex);
+    #endif
+
+    // Add and remove all the pending particle systems
+    for (const auto &particleSystem: particleSystemsToAdd)
+      particleSystems.emplace_back(particleSystem);
+    particleSystemsToAdd.clear();
+    for (const auto &particleSystem: particleSystemsToRemove)
+      std::erase(particleSystems, particleSystem);
+    particleSystemsToRemove.clear();
 
     glBlendFunc(GL_SRC_ALPHA, GL_ONE);
     for (const auto &particleSystem: particleSystems)
@@ -29,14 +45,10 @@ namespace Engine2D {
   }
 
   void ParticleSystemRenderer2D::addParticleSystem(const std::shared_ptr<ParticleSystem2D> &particleSystem) {
-    std::lock_guard lock(particleSystemMutex);
-    particleSystems.push_back(particleSystem);
+    particleSystemsToAdd.emplace_back(particleSystem);
   }
 
   void ParticleSystemRenderer2D::removeParticleSystem(const std::shared_ptr<ParticleSystem2D> &particleSystem) {
-    std::lock_guard lock(particleSystemMutex);
-    if (particleSystems.empty())
-      return;
-    std::erase(particleSystems, particleSystem);
+    particleSystemsToRemove.emplace_back(particleSystem);
   }
 } // Engine2D