+++ src/Engine2D/Physics/Physics2D.cpp
@@ -4,12 +4,8 @@
 // Date: 29/11/2024
 //
 
-#include <ranges>
-
 #include "Engine2D/Physics/Physics2D.hpp"
 #include "Engine2D/Entity2D.hpp"
-#include "Engine2D/Game2D.hpp"
-#include "Engine2D/Physics/Collider2D.hpp"
 #include "Engine2D/Physics/Collisions.hpp"
 #include "Engine2D/Physics/Rigidbody2D.hpp"
 #include "Engine/Settings.hpp"