+++ src/Engine2D/Transform2D.cpp
@@ -8,11 +8,8 @@
 #include <glm/gtx/rotate_vector.hpp>
 
 #include "Engine2D/Transform2D.hpp"
-#include "Engine2D/Entity2D.hpp"
 #include "Engine2D/Game2D.hpp"
 #include "Engine/Log.hpp"
-#include "Engine2D/Physics/Collisions.hpp"
-#include "Engine2D/Types/Vector2.hpp"
 
 namespace Engine2D {
   Transform2D::Transform2D(